+++ magick_mind/models/v2/reason.py
